Moved relaxed_combine_rows back into the main file.
Otherwise we end up with difficult circular dependencies between the
class definitions and the functions that need its type hints.
