+++ src/biocframe/__init__.py
@@ -15,6 +15,5 @@ except PackageNotFoundError:  # pragma: no cover
 finally:
     del version, PackageNotFoundError
 
-from .BiocFrame import BiocFrame, merge, relaxed_combine_columns
-from .relaxed_combine import relaxed_combine_rows
+from .BiocFrame import BiocFrame, relaxed_combine_rows, merge, relaxed_combine_columns
 from .io import from_pandas
